requirements.txt
email_listener
python-decouple
requests
asyncio
